Move checkers to their own executor file
