+++ executor/instruction/links_reverse_executor.rs
@@ -43,11 +43,12 @@ use crate::{
             tuple_relation_player_role_to_links_reverse, tuple_role_relation_player_to_links_reverse,
             unsafe_compare_result_tuple, TupleOrderingFn, TuplePositions,
         },
-        Checker, LinksIterateMode, VariableModes,
+        LinksIterateMode, VariableModes,
     },
     pipeline::stage::ExecutionContext,
     row::MaybeOwnedRow,
 };
+use crate::instruction::checker::Checker;
 
 pub(crate) struct LinksReverseExecutor {
     links: ir::pattern::constraint::Links<ExecutorVariable>,
